add final steps
