+++ src/judgeval/common/trainer/__init__.py
@@ -1,4 +1,5 @@
 from .trainer import JudgmentTrainer
 from .config import TrainerConfig
+from .trainable_model import TrainableModel
 
-__all__ = ["JudgmentTrainer", "TrainerConfig"]
+__all__ = ["JudgmentTrainer", "TrainerConfig", "TrainableModel"]
