+++ src/judgeval/common/trainer/trainer.py
@@ -1,105 +1,13 @@
 import asyncio
 import time
-from typing import Optional, Callable, Any, List
+from typing import Optional, Callable, Any, List, Union
 from fireworks import LLM, Dataset
 from .config import TrainerConfig
-from judgeval.tracer import Tracer, wrap
-
-
-class TrainableModel:
-    """
-    A wrapper class for managing model snapshots during training.
-
-    This class automatically handles model snapshot creation and management
-    during the GRPO (Generative Reinforcement Learning from Policy Optimization) process,
-    abstracting away manual snapshot management from users.
-    """
-
-    def __init__(self, config: TrainerConfig):
-        """
-        Initialize the TrainableModel.
-
-        Args:
-            config: TrainerConfig instance with model configuration
-        """
-        self.config = config
-        self.current_step = 0
-        self._current_model = None
-
-        # Initialize base model
-        self._base_model = self._create_base_model()
-        self._current_model = self._base_model
-
-    def _create_base_model(self):
-        """Create and configure the base model."""
-        base_model = LLM(
-            model=self.config.base_model_name,
-            deployment_type="on-demand",
-            id=self.config.deployment_id,
-            enable_addons=self.config.enable_addons,
-        )
-        base_model.apply()
-        return base_model
-
-    def get_current_model(self):
-        """Get the current model snapshot for generation."""
-        return self._current_model
-
-    @property
-    def chat(self):
-        """OpenAI-compatible chat interface."""
-        return self._current_model.chat
-
-    @property
-    def completions(self):
-        """OpenAI-compatible completions interface."""
-        return self._current_model.completions
-
-    def advance_to_next_step(self, step: int):
-        """
-        Advance to the next training step and update the current model snapshot.
-
-        Args:
-            step: The current training step number
-        """
-        self.current_step = step
-
-        if step == 0:
-            # Use base model for first step
-            self._current_model = self._base_model
-        else:
-            # Create new model snapshot from previous training step
-            model_name = (
-                f"accounts/{self.config.user_id}/models/{self.config.model_id}-v{step}"
-            )
-            self._current_model = LLM(
-                model=model_name,
-                deployment_type="on-demand-lora",
-                base_id=self.config.deployment_id,
-            )
-            # Ensure deployment is ready
-            self._current_model.apply()
-
-    def perform_reinforcement_step(self, dataset, step: int):
-        """
-        Perform a reinforcement learning step using the current model.
-
-        Args:
-            dataset: Training dataset for the reinforcement step
-            step: Current step number for output model naming
-
-        Returns:
-            Training job object
-        """
-        model_name = f"{self.config.model_id}-v{step + 1}"
-        return self._current_model.reinforcement_step(
-            dataset=dataset,
-            output_model=model_name,
-            epochs=self.config.epochs,
-            learning_rate=self.config.learning_rate,
-            accelerator_count=self.config.accelerator_count,
-            accelerator_type=self.config.accelerator_type,
-        )
+from .trainable_model import TrainableModel
+from judgeval.tracer import Tracer
+from judgeval.judgment_client import JudgmentClient
+from judgeval.scorers import BaseScorer, APIScorerConfig
+from judgeval.data import Example
 
 
 class JudgmentTrainer:
@@ -111,7 +19,11 @@ class JudgmentTrainer:
     """
 
     def __init__(
-        self, config: Optional[TrainerConfig] = None, tracer: Optional[Tracer] = None
+        self,
+        config: Optional[TrainerConfig] = None,
+        trainable_model: Optional[TrainableModel] = None,
+        tracer: Optional[Tracer] = None,
+        project_name: Optional[str] = None,
     ):
         """
         Initialize the JudgmentTrainer.
@@ -119,12 +31,18 @@ class JudgmentTrainer:
         Args:
             config: TrainerConfig instance with training parameters. If None, uses default config.
             tracer: Optional tracer for observability
+            trainable_model: Optional trainable model instance
+            project_name: Project name for organizing training runs and evaluations
         """
         self.config = config or TrainerConfig()
         self.tracer = tracer
+        self.project_name = project_name or "judgment_training"
 
         # Initialize trainable model wrapper
-        self.trainable_model = wrap(TrainableModel(self.config))
+        self.trainable_model = trainable_model
+
+        # Initialize judgment client for evaluation
+        self.judgment_client = JudgmentClient()
 
     def _initialize_base_model(self):
         """Initialize the base model with PEFT addon support."""
@@ -143,7 +61,7 @@ class JudgmentTrainer:
     async def generate_rollouts_and_rewards(
         self,
         agent_function: Callable[[Any], Any],
-        reward_function: Callable[[Any, Any], float],
+        scorers: List[Union[APIScorerConfig, BaseScorer]],
         prompts: List[Any],
         num_prompts: Optional[int] = None,
         num_generations_per_prompt: Optional[int] = None,
@@ -155,7 +73,7 @@ class JudgmentTrainer:
 
         Args:
             agent_function: Function/agent to call for generating responses
-            reward_function: Function to compute reward given prompt and response
+            scorers: List of scorer objects to evaluate responses
             prompts: List of prompts to use for training
             num_prompts: Number of prompts to use (defaults to config value, limited by prompts list length)
             num_generations_per_prompt: Generations per prompt (defaults to config value)
@@ -180,11 +98,7 @@ class JudgmentTrainer:
                 prompt_input = prompts[prompt_id]
 
                 # Call the agent function with the current model and prompt
-                response_data = await agent_function(
-                    model=self.trainable_model,
-                    prompt_input=prompt_input,
-                    config=self.config,
-                )
+                response_data = await agent_function(prompt_input)
 
                 # Extract messages from response_data or trace
                 messages = response_data.get("messages", [])
@@ -200,8 +114,29 @@ class JudgmentTrainer:
                     # Fallback to response_data messages if trace extraction fails
                     pass
 
-                # Compute reward using provided reward function
-                reward = reward_function(prompt_input, response_data)
+                # Create an Example object from the response data for evaluation
+                # Include prompt_input, messages, and response_data as requested
+                example = Example(
+                    input=prompt_input, messages=messages, actual_output=response_data
+                )
+
+                # Use run_evaluation to compute reward using scorer objects
+                scoring_results = self.judgment_client.run_evaluation(
+                    examples=[example],
+                    scorers=scorers,
+                    project_name=self.project_name,
+                    eval_run_name=f"training_step_{self.trainable_model.current_step}_prompt_{prompt_id}_gen_{generation_id}",
+                )
+
+                # Extract reward from scoring results
+                # Take the average score across all scorers as the reward
+                if scoring_results and scoring_results[0].scorers_data:
+                    reward = sum(
+                        scorer_data.score
+                        for scorer_data in scoring_results[0].scorers_data
+                    ) / len(scoring_results[0].scorers_data)
+                else:
+                    reward = 0.0
 
             return {
                 "prompt_id": prompt_id,
@@ -244,7 +179,7 @@ class JudgmentTrainer:
     async def run_reinforcement_learning(
         self,
         agent_function: Callable[[Any], Any],
-        reward_function: Callable[[Any, Any], float],
+        scorers: List[Union[APIScorerConfig, BaseScorer]],
         prompts: List[Any],
     ):
         """
@@ -252,13 +187,13 @@ class JudgmentTrainer:
 
         This method performs multiple steps of reinforcement learning, where each step:
         1. Advances to the appropriate model snapshot
-        2. Generates rollouts and computes rewards
+        2. Generates rollouts and computes rewards using scorers
         3. Trains a new model using reinforcement learning
         4. Waits for training completion
 
         Args:
             agent_function: Function/agent to call for generating responses
-            reward_function: Function to compute reward given prompt and response
+            scorers: List of scorer objects to evaluate responses
             prompts: List of prompts to use for training
         """
 
@@ -273,7 +208,7 @@ class JudgmentTrainer:
 
             # Generate rollouts and rewards using current model snapshot
             dataset_rows = await self.generate_rollouts_and_rewards(
-                agent_function, reward_function, prompts
+                agent_function, scorers, prompts
             )
 
             # Create dataset from dataset rows
@@ -302,7 +237,7 @@ class JudgmentTrainer:
     def train(
         self,
         agent_function: Callable[[Any], Any],
-        reward_function: Callable[[Any, Any], float],
+        scorers: List[Union[APIScorerConfig, BaseScorer]],
         prompts: List[Any],
     ):
         """
@@ -313,9 +248,7 @@ class JudgmentTrainer:
         Args:
             agent_function: Function/agent to call for generating responses.
                            Should accept (model, prompt_input, config) and return response data
-            reward_function: Function to compute reward given (prompt_input, response_data)
+            scorers: List of scorer objects to evaluate responses
             prompts: List of prompts to use for training
         """
-        asyncio.run(
-            self.run_reinforcement_learning(agent_function, reward_function, prompts)
-        )
+        asyncio.run(self.run_reinforcement_learning(agent_function, scorers, prompts))
